Renamed npmi filter package.
