+++ service/src/test/java/org/dice_research/fc/paths/search/SPARQLBasedSOPathSearcherTest.java
@@ -15,9 +15,9 @@ import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
 import org.apache.jena.rdf.model.Property;
 import org.apache.jena.rdf.model.Resource;
-import org.dice.FactCheck.Corraborative.filter.npmi.NPMIFilterException;
 import org.dice_research.fc.data.Predicate;
 import org.dice_research.fc.data.QRestrictedPath;
+import org.dice_research.fc.paths.scorer.filter.NPMIFilterException;
 import org.dice_research.fc.sparql.filter.EqualsFilter;
 import org.dice_research.fc.sparql.filter.IRIFilter;
 import org.dice_research.fc.sparql.filter.NamespaceFilter;
